[TEST] View, Form, Template files were created
Form file was created to render the document form to the view. Images were added to the form field. The template file was created to test render the form. After rendering the form, it was able to extract 1 pdf file from multiple images.
To convert this multiple images 2 additional modules were used.
tempfile - Built-in
fpdf - pypi
tempfile was used to store the multiple images in a temporary location.
fpdf was used to merge this images in to a single pdf file.